+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project uses [Semantic Versioning](https://semver.org/). Major versions are
 
 ## [Unreleased]
 
+### Changed
+- Opened 1.0.6-dev development cycle.
+
 ## [1.0.5] - 2026-06-27
 
 ### Fixed